@@ -1011,12 +1011,12 @@ const HLEFunction UtilsForKernel[] = {
 	{0X6C6887EE, nullptr,                                            "UtilsForKernel_6C6887EE",                   '?', ""        },
 	{0X91E4F6A7, nullptr,                                            "sceKernelLibcClock",                        '?', ""        },
 	{0X27CC57F0, nullptr,                                            "sceKernelLibcTime",                         '?', ""        },
-	{0X79D1C3FA, nullptr,                                            "sceKernelDcacheWritebackAll",               '?', ""        },
-	{0X3EE30821, nullptr,                                            "sceKernelDcacheWritebackRange",             '?', ""        },
-	{0X34B9FA9E, nullptr,                                            "sceKernelDcacheWritebackInvalidateRange",   '?', ""        },
-	{0XB435DEC5, nullptr,                                            "sceKernelDcacheWritebackInvalidateAll",     '?', ""        },
-	{0XBFA98062, nullptr,                                            "sceKernelDcacheInvalidateRange",            '?', ""        },
-	{0X920F104A, nullptr,                                            "sceKernelIcacheInvalidateAll",              '?', ""        },
+	{0X79D1C3FA, &WrapI_V<sceKernelDcacheWritebackAll>,              "sceKernelDcacheWritebackAll",               '?', ""        },
+	{0X3EE30821, &WrapI_UI<sceKernelDcacheWritebackRange>,           "sceKernelDcacheWritebackRange",             '?', ""        },
+	{0X34B9FA9E, &WrapI_UI<sceKernelDcacheWritebackInvalidateRange>, "sceKernelDcacheWritebackInvalidateRange",   '?', ""        },
+	{0XB435DEC5, &WrapI_V<sceKernelDcacheWritebackInvalidateAll>,    "sceKernelDcacheWritebackInvalidateAll",     '?', ""        },
+	{0XBFA98062, &WrapI_UI<sceKernelDcacheInvalidateRange>,          "sceKernelDcacheInvalidateRange",            '?', ""        },
+	{0X920F104A, &WrapU_V<sceKernelIcacheInvalidateAll>,             "sceKernelIcacheInvalidateAll",              '?', ""        },
 	{0XE860E75E, nullptr,                                            "sceKernelUtilsMt19937Init",                 '?', ""        },
 	{0X06FB8A63, nullptr,                                            "sceKernelUtilsMt19937UInt",                 '?', ""        },
 };
@@ -1416,6 +1416,7 @@ const char *KernelErrorToString(u32 err) {
 	case SCE_ERROR_ATRAC_IS_FOR_SCESAS: return "SCE_ERROR_ATRAC_IS_FOR_SCESAS";
 	case SCE_ERROR_ATRAC_AA3_INVALID_DATA: return "SCE_ERROR_ATRAC_AA3_INVALID_DATA";
 	case SCE_ERROR_ATRAC_AA3_SIZE_TOO_SMALL: return "SCE_ERROR_ATRAC_AA3_SIZE_TOO_SMALL";
+	case SCE_ERROR_ATRAC_AA3_BAD_CODEC_PARAMS: return "SCE_ERROR_ATRAC_AA3_BAD_CODEC_PARAMS";
 	case SCE_ERROR_ATRAC_BAD_ALIGNMENT: return "SCE_ERROR_ATRAC_BAD_ALIGNMENT";
 
 	case SCE_MPEG_ERROR_BAD_VERSION: return "SCE_MPEG_ERROR_BAD_VERSION";
