@@ -290,15 +290,25 @@ void ImGui_ImplThin3d_DestroyDeviceObjects() {
 	}
 }
 
-bool ImGui_ImplThin3d_Init(Draw::DrawContext *draw, const uint8_t *ttf_font, size_t size) {
+bool ImGui_ImplThin3d_Init(Draw::DrawContext *draw,
+	const uint8_t *ttf_font_proportional, size_t proportional_size,
+	const uint8_t *ttf_font_fixed, size_t fixed_size) {
 	ImGuiIO& io = ImGui::GetIO();
-	if (ttf_font) {
-		g_proportionalFont = io.Fonts->AddFontFromMemoryTTF((void *)ttf_font, (int)size, 21.0f / g_display.dpi_scale_x, nullptr, io.Fonts->GetGlyphRangesDefault());
+	g_proportionalFont = nullptr;
+	if (ttf_font_proportional) {
+		g_proportionalFont = io.Fonts->AddFontFromMemoryTTF((void *)ttf_font_proportional, (int)proportional_size, 21.0f / g_display.dpi_scale_x, nullptr, io.Fonts->GetGlyphRangesDefault());
+	}
+	if (ttf_font_fixed) {
+		g_fixedFont = io.Fonts->AddFontFromMemoryTTF((void *)ttf_font_fixed, (int)fixed_size, 20.0f / g_display.dpi_scale_x, nullptr, io.Fonts->GetGlyphRangesDefault());
 	} else {
-		// fallback
+		g_fixedFont = io.Fonts->AddFontDefault();
+	}
+
+	if (!g_proportionalFont) {
 		g_proportionalFont = g_fixedFont;
 	}
-	g_fixedFont = io.Fonts->AddFontDefault();
+
+	// g_fixedFont = io.Fonts->AddFontDefault();
 	ImGui::GetStyle().ScaleAllSizes(1.0f / g_display.dpi_scale_x);
 	ImGui::GetStyle().Colors[ImGuiCol_Border] = ImColor(IM_COL32(0x2A, 0x2F, 0x3B, 0xFF));
 
